# Initial Commit. The Bot is working and tested to some extend.
Get your Telegram Bot Token from the Bot-Father and paste it as the first line in "telegramtoken".
Type your pr0 username in the first and your password in the second line of the file "pr0token".

The provided Shitt0lk.service is a very basic service I use with systemd.
You can build your service upon it.